minor refactoring + missing translations
