+++ README.md
@@ -105,9 +105,9 @@ The app is currently available in the following languages:
 * French
 * German
 * Korean
+* Portuguese
 * Russian
 * Spanish
-* Portuguese (Brazil)
 
 More translations are welcome. See [instructions](https://github.com/tinode/chat/blob/devel/docs/translations.md#android).
