+++ app/src/main/java/co/tinode/tindroid/MessagesAdapter.java
@@ -107,8 +107,7 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.ViewHo
     private static final int REFRESH_HARD = 2;
 
     // Bits defining message bubble variations
-    // _TIP_ == "single", i.e. has a bubble tip.
-    // _SIMPLE_ == no bubble tip, just a rounded rectangle.
+    // _TIP == "single", i.e. has a bubble tip.
     // _DATE == the date bubble is visible.
     private static final int VIEWTYPE_SIDE_LEFT   = 0b000010;
     private static final int VIEWTYPE_SIDE_RIGHT  = 0b000100;
@@ -139,7 +138,7 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.ViewHo
     private SparseBooleanArray mSelectedItems = null;
     private int mPagesToLoad;
 
-    private final MediaControl mMediaControl = new MediaControl();
+    private final MediaControl mMediaControl;
 
     MessagesAdapter(@NonNull MessageActivity context, @NonNull SwipeRefreshLayout refresher) {
         super();
@@ -151,6 +150,8 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.ViewHo
         mPagesToLoad = 1;
 
         mMessageLoaderCallback = new MessageLoaderCallbacks();
+
+        mMediaControl = new MediaControl(context);
 
         mSelectionModeCallback = new ActionMode.Callback() {
             @Override
@@ -1337,256 +1338,6 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.ViewHo
             mActivity.showFragment(MessageActivity.FRAGMENT_VIEW_VIDEO, args, true);
 
             return true;
-        }
-    }
-
-    // Actions to take in setOnPreparedListener, when the player is ready.
-    private enum PlayerReadyAction {
-        // Do nothing.
-        NOOP,
-        // Start playing.
-        PLAY,
-        // Seek without changing player state.
-        SEEK,
-        // Seek, then play when seek finishes.
-        SEEKNPLAY
-    }
-
-    private class MediaControl {
-        private AudioManager mAudioManager = null;
-        private MediaPlayer mAudioPlayer = null;
-        private int mPlayingAudioSeq = -1;
-        private FullFormatter.AudioControlCallback mAudioControlCallback = null;
-        // Action to take when the player becomes ready.
-        private PlayerReadyAction mReadyAction = PlayerReadyAction.NOOP;
-        // Playback fraction to seek to when the player is ready.
-        private float mSeekTo = -1f;
-
-        boolean ensurePlayerReady(final int seq, Map<String, Object> data,
-                                       FullFormatter.AudioControlCallback control) throws IOException {
-            if (mAudioPlayer != null && mPlayingAudioSeq == seq) {
-                mAudioControlCallback = control;
-                return true;
-            }
-
-            if (mPlayingAudioSeq > 0 && mAudioControlCallback != null) {
-                mAudioControlCallback.reset();
-            }
-
-            // Declare current player un-prepared.
-            mPlayingAudioSeq = -1;
-
-            if (mAudioPlayer != null) {
-                try {
-                    mAudioPlayer.stop();
-                } catch (IllegalStateException ignored) {}
-                mAudioPlayer.reset();
-            } else {
-                mAudioPlayer = new MediaPlayer();
-            }
-
-            if (mAudioManager == null) {
-                mAudioManager = (AudioManager) mActivity.getSystemService(Activity.AUDIO_SERVICE);
-                mAudioManager.setMode(AudioManager.MODE_IN_CALL);
-                mAudioManager.setSpeakerphoneOn(true);
-            }
-
-            mAudioControlCallback = control;
-            mAudioPlayer.setOnPreparedListener(mp -> {
-                if (mPlayingAudioSeq > 0) {
-                    // Another media have already been started while we waited for this media to become ready.
-                    mp.release();
-                    return;
-                }
-
-                mPlayingAudioSeq = seq;
-                if (mReadyAction == PlayerReadyAction.PLAY) {
-                    mReadyAction = PlayerReadyAction.NOOP;
-                    mp.start();
-                } else if (mReadyAction == PlayerReadyAction.SEEK ||
-                        mReadyAction == PlayerReadyAction.SEEKNPLAY) {
-                    seekTo(fractionToPos(mSeekTo));
-                }
-                mSeekTo = -1f;
-            });
-            mAudioPlayer.setOnCompletionListener(mp -> {
-                if (mPlayingAudioSeq != seq) {
-                    return;
-                }
-                int pos = mp.getCurrentPosition();
-                if (pos > 0) {
-                    if (mAudioControlCallback != null) {
-                        mAudioControlCallback.reset();
-                    }
-                }
-            });
-            mAudioPlayer.setOnErrorListener((mp, what, extra) -> {
-                Log.w(TAG, "Playback error " + what + "/" + extra);
-                if (mPlayingAudioSeq != seq) {
-                    return true;
-                }
-                Toast.makeText(mActivity, R.string.unable_to_play_audio, Toast.LENGTH_SHORT).show();
-                return false;
-            });
-            mAudioPlayer.setOnSeekCompleteListener(mp -> {
-                if (mPlayingAudioSeq != seq) {
-                    return;
-                }
-                if (mReadyAction == PlayerReadyAction.SEEKNPLAY) {
-                    mReadyAction = PlayerReadyAction.NOOP;
-                    mp.start();
-                }
-            });
-            mAudioPlayer.setAudioAttributes(
-                    new AudioAttributes.Builder()
-                            .setLegacyStreamType(AudioManager.STREAM_VOICE_CALL).build());
-
-            Object val;
-            if ((val = data.get("ref")) instanceof String) {
-                Tinode tinode = Cache.getTinode();
-                URL url = tinode.toAbsoluteURL((String) val);
-                if (url != null) {
-                    try {
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                            mAudioPlayer.setDataSource(mActivity, Uri.parse(url.toString()),
-                                    tinode.getRequestHeaders(), null);
-                        } else {
-                            Uri uri = Uri.parse(url.toString()).buildUpon()
-                                    .appendQueryParameter("apikey", tinode.getApiKey())
-                                    .appendQueryParameter("auth", "token")
-                                    .appendQueryParameter("secret", tinode.getAuthToken())
-                                    .build();
-                            mAudioPlayer.setDataSource(mActivity, uri);
-                        }
-                    } catch (SecurityException | IOException ex) {
-                        Log.w(TAG, "Failed to add URI data source ", ex);
-                        Toast.makeText(mActivity, R.string.unable_to_play_audio, Toast.LENGTH_SHORT).show();
-                    }
-                } else {
-                    mAudioControlCallback.reset();
-                    Log.w(TAG, "Invalid ref URL " + val);
-                    Toast.makeText(mActivity, R.string.unable_to_play_audio, Toast.LENGTH_SHORT).show();
-                    return false;
-                }
-            } else if ((val = data.get("val")) instanceof String) {
-                byte[] source = Base64.decode((String) val, Base64.DEFAULT);
-                mAudioPlayer.setDataSource(new MemoryAudioSource(source));
-            } else {
-                mAudioControlCallback.reset();
-                Log.w(TAG, "Unable to play audio: missing data");
-                Toast.makeText(mActivity, R.string.unable_to_play_audio, Toast.LENGTH_SHORT).show();
-                return false;
-            }
-            mAudioPlayer.prepareAsync();
-            return true;
-        }
-
-        void releasePlayer(int seq) {
-            if ((seq != 0 && mPlayingAudioSeq != seq) || mPlayingAudioSeq == -1) {
-                return;
-            }
-
-            mPlayingAudioSeq = -1;
-            mReadyAction = PlayerReadyAction.NOOP;
-            mSeekTo = -1f;
-            if (mAudioPlayer != null) {
-                try {
-                    mAudioPlayer.stop();
-                } catch (IllegalStateException ignored) {}
-                mAudioPlayer.reset();
-                mAudioPlayer.release();
-                mAudioPlayer = null;
-            }
-            if (mAudioControlCallback != null) {
-                mAudioControlCallback.reset();
-            }
-        }
-
-        // Start playing at the current position.
-        void playWhenReady() {
-            if (mPlayingAudioSeq > 0) {
-                mAudioPlayer.start();
-            } else {
-                mReadyAction = PlayerReadyAction.PLAY;
-            }
-        }
-
-        void pause() {
-            if (mAudioPlayer != null && mAudioPlayer.isPlaying()) {
-                mAudioPlayer.pause();
-            }
-            mReadyAction = PlayerReadyAction.NOOP;
-            mSeekTo = -1f;
-        }
-
-        void seekToWhenReady(float fraction) {
-            if (mPlayingAudioSeq > 0) {
-                // Already prepared.
-                int pos = fractionToPos(fraction);
-                if (mAudioPlayer.getCurrentPosition() != pos) {
-                    // Need to seek.
-                    mReadyAction = PlayerReadyAction.NOOP;
-                    seekTo(pos);
-                } else {
-                    // Already prepared & at the right position.
-                    mAudioPlayer.start();
-                }
-            } else {
-                mReadyAction = PlayerReadyAction.SEEK;
-                mSeekTo = fraction;
-            }
-        }
-
-        void seekTo(int pos) {
-            if (mAudioPlayer != null && mPlayingAudioSeq > 0) {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                    mAudioPlayer.seekTo(pos, MediaPlayer.SEEK_CLOSEST);
-                } else {
-                    mAudioPlayer.seekTo(pos);
-                }
-            }
-        }
-
-        private int fractionToPos(float fraction) {
-            try {
-                if (mAudioPlayer != null && mPlayingAudioSeq > 0) {
-                    long duration = mAudioPlayer.getDuration();
-                    if (duration > 0) {
-                        return ((int) (fraction * duration));
-                    } else {
-                        Log.w(TAG, "Audio has no duration");
-                    }
-                }
-            } catch (IllegalStateException ex) {
-                Log.w(TAG, "Duration not available yet " + mPlayingAudioSeq, ex);
-            }
-            return -1;
-        }
-    }
-
-    // Wrap in-band audio into MediaDataSource to make it playable by MediaPlayer.
-    private static class MemoryAudioSource extends MediaDataSource {
-        private final byte[] mData;
-
-        MemoryAudioSource(byte[] source) {
-            mData = source;
-        }
-
-        @Override
-        public int readAt(long position, byte[] destination, int offset, int size) throws IOException {
-            size = Math.min(mData.length - (int) position, size);
-            System.arraycopy(mData, (int) position, destination, offset, size);
-            return size;
-        }
-
-        @Override
-        public long getSize() throws IOException {
-            return mData.length;
-        }
-
-        @Override
-        public void close() throws IOException {
-            // Do nothing.
         }
     }
 }
