+++ app/src/main/res/values-ru/strings.xml
@@ -311,4 +311,9 @@
     <string name="unable_to_play_video">Невозможно показать это видео</string>
     <string name="select_image_or_video">Выбрать фото или видео</string>
     <string name="tap_to_select">Нажмите чтобы выбрать</string>
+    <string name="video_call_channel_name">Видеозвонки</string>
+    <string name="video_call_channel_description">Видео- и аудиовызовы</string>
+    <string name="tinode_video_call">Видеозвонок Tinode</string>
+    <string name="answer_call">Ответить</string>
+    <string name="decline_call">Отклонить</string>
 </resources>
