+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -331,4 +331,9 @@
     <string name="unable_to_play_video">無法播放此視頻。</string>
     <string name="select_image_or_video">選擇圖片或視頻</string>
     <string name="tap_to_select">點擊選擇</string>
+    <string name="video_call_channel_name">視頻通話</string>
+    <string name="video_call_channel_description">視頻和音頻通話</string>
+    <string name="tinode_video_call">Tinode 視頻通話</string>
+    <string name="answer_call">接聽</string>
+    <string name="decline_call">拒絕</string>
 </resources>
