+++ app/src/main/res/values-zh/strings.xml
@@ -332,4 +332,9 @@
     <string name="unable_to_play_video">无法播放此视频。</string>
     <string name="select_image_or_video">选择图片或视频</string>
     <string name="tap_to_select">点击选择</string>
+    <string name="video_call_channel_name">视频通话</string>
+    <string name="video_call_channel_description">视频和音频通话</string>
+    <string name="tinode_video_call">Tinode 视频通话</string>
+    <string name="answer_call">接听</string>
+    <string name="decline_call">拒绝</string>
 </resources>
